+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [0.75.0]
 
+* Add support for renaming variables within the scope of a predicate clause, grammar rule, or directive
 * Add "Unify with new variable" refactoring support for selected terms in predicate rules and grammar rules
 * Add "Inline variable" refactoring support for replacing variable unification goals in predicate rules and grammar rules
 * Add "Wrap file contents as an object" refactoring support for converting plain Prolog files to Logtalk objects